+++ examples/image.rs
@@ -26,7 +26,11 @@
 #![no_main]
 
 use cortex_m_rt::{entry, exception, ExceptionFrame};
-use embedded_graphics::{image::ImageLE, prelude::*};
+use embedded_graphics::{
+    image::{Image, ImageRawLE},
+    pixelcolor::Rgb565,
+    prelude::*,
+};
 use panic_semihosting as _;
 use ssd1331::{DisplayRotation::Rotate0, Ssd1331};
 use stm32f1xx_hal::{
@@ -82,10 +86,11 @@ fn main() -> ! {
 
     // Loads an 86x64px image encoded in LE (Little Endian) format. This image is a 16BPP image of
     // the Rust mascot, Ferris.
-    let im = ImageLE::new(include_bytes!("./ferris.raw"), 86, 64)
-        .translate(Point::new((96 - 86) / 2, 0));
+    let raw: ImageRawLE<Rgb565> = ImageRawLE::new(include_bytes!("./ferris.raw"), 86, 64);
 
-    im.draw(&mut disp);
+    Image::new(&raw, Point::new((96 - 86) / 2, 0))
+        .draw(&mut disp)
+        .unwrap();
 
     disp.flush().unwrap();