+++ src/display.rs
@@ -369,20 +369,25 @@ where
     SPI: hal::blocking::spi::Write<u8>,
     DC: OutputPin,
 {
-    fn draw_pixel(&mut self, pixel: drawable::Pixel<Rgb565>) {
+    type Error = core::convert::Infallible;
+
+    fn draw_pixel(&mut self, pixel: drawable::Pixel<Rgb565>) -> Result<(), Self::Error> {
         let drawable::Pixel(pos, color) = pixel;
 
         // Guard against negative values. All positive i32 values from `pos` can be represented in
-        // the `u32`s that `set_pixel()` accepts.
+        // the `u32`s that `set_pixel()` accepts...
         if pos.x < 0 || pos.y < 0 {
-            return;
+            return Ok(());
         }
 
+        // ... which makes the `as` coercions here safe.
         self.set_pixel(
-            (pos.x).try_into().unwrap(),
-            (pos.y).try_into().unwrap(),
+            (pos.x) as u32,
+            (pos.y) as u32,
             RawU16::from(color).into_inner(),
         );
+
+        Ok(())
     }
 
     fn size(&self) -> Size {